+++ CotEditor/Sources/FindPanelButtonViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2020 1024jp
+//  © 2014-2022 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -23,14 +23,11 @@
 //  limitations under the License.
 //
 
-import Combine
 import Cocoa
 
 final class FindPanelButtonViewController: NSViewController {
     
     // MARK: Private Properties
-    
-    private var findNextAfterReplaceObserver: AnyCancellable?
     
     @IBOutlet private weak var findAllButton: NSButton?
     @IBOutlet private weak var replaceButton: NSButton?
@@ -47,32 +44,14 @@ final class FindPanelButtonViewController: NSViewController {
         // workaround an issue that NSComboButton cannnot be localized by .strings file
         self.findAllButton?.title = NSLocalizedString("Find All", comment: "")
         
-        // change "Replace" button behavior depending on the user setting
-        self.findNextAfterReplaceObserver = UserDefaults.standard.publisher(for: .findNextAfterReplace, initial: true)
-            .map { $0
-                ? "Replace the current selection with the replacement text, then find the next match.".localized
-                : "Replace the current selection with the replacement text.".localized
-            }
-            .assign(to: \.toolTip, on: self.replaceButton!)
+        self.replaceButton?.toolTip = "Replace the current selection with the replacement text, then find the next match.".localized
     }
     
     
     
     // MARK: Action Messages
     
-    /// replace next matched string with given string
-    @IBAction func replace(_ sender: Any?) {
-        
-        // perform "Replace & Find" instead of "Replace"
-        if UserDefaults.standard[.findNextAfterReplace] {
-            TextFinder.shared.replaceAndFind(sender)
-        } else {
-            TextFinder.shared.replace(sender)
-        }
-    }
-    
-    
-    /// perform segmented Find Next/Previous button
+    /// Perform the segmented Find Next/Previous button.
     @IBAction func clickSegmentedFindButton(_ sender: NSSegmentedControl) {
         
         switch sender.selectedSegment {