+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Change Log
 - Display the concrete progress of the find/replace task in the progress dialog.
 - Update the Unicode block name list for the character inspector from Unicode 14.0.0 to Unicode 15.0.0.
 - Deprecate the option to keep the progress dialog for find/replacement after finishing the task.
+- Deprecate the option not to select the next match after the replacement when clicking the Replace button. 
 - [trivial] Tweak the style of headings in menus.
 - [trivial] Tweak the setting summary display in the print dialog.
 - [dev] Migrate the most of sheets and popovers to SwiftUI.